+++ releases.md
@@ -1,5 +1,9 @@
 # Releases
 
+## Unreleased
+
+  - Fix handling of inspect on Ruby 4+.
+
 ## v0.3.1
 
   - `Client#run` now returns an `Async::Task` (as it did in earlier releases).